+++ config/node/on_create_node.js
@@ -2,7 +2,7 @@ const path = require("path")
 const isString = require("lodash/isString")
 const isUndefined = require("lodash/isUndefined")
 
-const { isURL } = require("../../src/utils/url_utils")
+const { isURL } = require("../../src/utils/urlUtils")
 
 const { NODE_ENV } = process.env
