+++ src/components/Home/HeroSection/Title.js
@@ -1,7 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
 
-import useDetectJavascript from "src/utils/use_detect_javascript"
+import useDetectJavascript from "src/utils/useDetectJavascript"
 import AnimatedLetters from "./Title/AnimatedLetters"
 
 import styles from "./Title.module.scss"
