+++ src/components/Home/VenturesSection.js
@@ -1,10 +1,10 @@
 import React from "react"
 import PropTypes from "prop-types"
 
-import CallToAction from "src/components/call_to_action"
-import SectionTitle from "src/components/section_title"
-import SubsectionTitle from "src/components/subsection_title"
-import Text from "src/components/text"
+import CallToAction from "src/components/CallToAction"
+import SectionTitle from "src/components/SectionTitle"
+import SubsectionTitle from "src/components/SubsectionTitle"
+import Text from "src/components/Text"
 import HelpingWithSubsection from "./VenturesSection/HelpingWithSubsection"
 import LookingForSubsection from "./VenturesSection/LookingForSubsection"
 import PortfolioSubsection from "./VenturesSection/PortfolioSubsection"
